requirements.txt
betterproto[compiler]
ruamel.yaml